完善后续功能

+++ blogserver/src/main/java/com/sf/blogserver/service/impl/CategoryServiceImpl.java
@@ -1,25 +1,24 @@
 package com.sf.blogserver.service.impl;
 
 import com.sf.blogserver.bean.Category;
+import com.sf.blogserver.mapper.ArticleMapper;
 import com.sf.blogserver.mapper.CategoryMapper;
+import com.sf.blogserver.mapper.IssueMapper;
 import com.sf.blogserver.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-/**
- * @Discription
- * @auther Hh
- * @package com.sf.blogserver.service.impl
- * @create 2019/10/28 8:50
- * @Version: 1.0
- */
 @Service
 public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     CategoryMapper categoryMapper;
+    @Autowired
+    ArticleMapper articleMapper;
+    @Autowired
+    IssueMapper issueMapper;
 
     @Override
     public List<Category> getAllCategory() {
@@ -33,6 +32,14 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public int deleteCategoryByPrimaryKey(Integer categoryId) {
+        if(articleMapper.selectByCategoryId(categoryId).size() > 0 || issueMapper.selectByCategoryId(categoryId).size() > 0){
+            return -1;
+        }
         return categoryMapper.deleteByPrimaryKey(categoryId);
+    }
+
+    @Override
+    public int update(Category category) {
+        return categoryMapper.updateByPrimaryKeySelective(category);
     }
 }
